requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-decouple==3.8
sqlparse==0.5.3
